Updated to 2.80(maybe?)
I ran your add-on through my updater add-on, there might still be things which needs an update, but at least it now starts without error messages and it seems like the UI is working(3D View > Sidebar). I never used you add-on before, so I neither don't know how to use it, or if it is working as intended.  

I would like to try your add-on, because I'm working on an add-on which can extract information from a screenplay(written in Blender), and then auto add scenes and dialogue in the timing of the screenplay into the VSE. A nice thing would be if it was possible to assign 3D assets to specific words in the screenplay for like locations, props and characters, and then auto add them to the scenes they're mentioned in. Or in other words by the click of a button you could convert your screenplay into a film, consisting of 3d scenes with the assets needed in that scene, ready for animation and camera blocking. 

And my though was that maybe your tool could have features added, so this would become possible?